+++ src/Component/Power/InitPcu_InitialStudy.cpp
@@ -9,7 +9,7 @@
 PCU_InitialStudy InitPCU_InitialStudy(ClockGenerator* clock_gen, int pcu_id, const std::string fname, const std::vector<SAP*> saps, BAT* bat, double compo_step_time) {
   IniAccess pcu_conf(fname);
 
-  const std::string st_pcu_id = std::to_string(static_cast<long long>(pcu_id));
+  const std::string st_pcu_id = std::to_string(pcu_id);
   const char* cs = st_pcu_id.data();
 
   char Section[30] = "PCU";
